generator.py: Run macros on headers, too, to embed images, see:
Sorry that I am not familiar with Python, so I'm not sure if this is a
correct fix. At least this works for me, I guess. The problem is that
I want to put images on headers, some because then it could be put on
the center of the slide, and some because I just want an image on
a header. I found that headers are skipped to be processed by macros,
so that images on headers aren't embedded. I am not sure if this is
designed to work this way, but it works for me... :)

Thank you for reading and considering!