Now if a theme are using the @import this file are embeded too.
Note: If the original theme are using images and you want to use those
images in the new theme, this images must be copied to the new theme or
linked (with ln -s).

I've solved a minimal bug: when a user uses the -i option and the image
is not on the correct place the html.replace tries to replace with a
boolean value (which gives a coercing to unicode error).